% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VAR_OLS.R
\name{impulse_response}
\alias{impulse_response}
\alias{impulse_response.default}
\alias{impulse_response.OLS}
\title{Impulse-Response function of estimated VAR model
provide an impulse-response function and its plot}
\usage{
impulse_response(
  object,
  p,
  variable,
  period,
  type = c("origin", "structural"),
  ncol.fig = 2,
  ...
)

\method{impulse_response}{default}(
  object,
  p,
  variable,
  period,
  type = c("origin", "structural"),
  ncol.fig = 2
)

\method{impulse_response}{OLS}(
  object,
  p,
  variable,
  period,
  type = c("origin", "structural"),
  ncol.fig = 2
)
}
\arguments{
\item{object}{a model object}

\item{p}{a integer, time lags of endogenous variables}

\item{variable}{a character, a variable which gives a shock}

\item{period}{a integer, time period}

\item{type}{shock to be used, \code{structural} stands for independent shocks.}

\item{ncol.fig}{a integer, number of figures plotted in a row}

\item{...}{extra arguments}
}
\description{
Impulse-Response function of estimated VAR model
provide an impulse-response function and its plot
}
\section{Methods (by class)}{
\itemize{
\item \code{default}: default method

\item \code{OLS}: impulse-response function of a OLS VAR model
}}

