% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayesVAR.R
\name{plot.bayesVAR}
\alias{plot.bayesVAR}
\title{Figures from Bayesian VAR model
provide figures from Bayesian VAR methods}
\usage{
\method{plot}{bayesVAR}(model, ncol.fig = 2, type = c("fitting", "forecast"))
}
\arguments{
\item{model}{Bayesian VAR model}

\item{ncol.fig}{a integer, number of figures plotted in a row}

\item{type}{plotting for goodness-of-fitted or forecasting}
}
\description{
Figures from Bayesian VAR model
provide figures from Bayesian VAR methods
}
