% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VAR_OLS.R
\name{plot.OLS}
\alias{plot.OLS}
\title{Figures from OLS VAR models
provide figures from OLS VAR methods}
\usage{
\method{plot}{OLS}(model, ncol.fig = 2)
}
\arguments{
\item{model}{OLS VAR model}

\item{ncol.fig}{a integer, number of figures plotted in a row}
}
\description{
Figures from OLS VAR models
provide figures from OLS VAR methods
}
