% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayesVAR.R
\name{predict.bayesVAR}
\alias{predict.bayesVAR}
\title{Posterior predictive distributions from Bayesian VAR model
sampling posterior predictive distributions from the Bayesian VAR model}
\usage{
\method{predict}{bayesVAR}(object, newdata, condition, type = c("unconditional", "conditional"))
}
\arguments{
\item{object}{Bayesian VAR model}

\item{newdata}{a data of xts class, exogenous variables for the prediction period}

\item{condition}{a data of xts class, endogenous variables to be conditioned}

\item{type}{plotting for goodness-of-fitted or forecasting}
}
\description{
Posterior predictive distributions from Bayesian VAR model
sampling posterior predictive distributions from the Bayesian VAR model
}
