% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VAR_OLS.R
\name{predict_simple}
\alias{predict_simple}
\title{Simple point predicition with OLS parameter estimates}
\usage{
predict_simple(Y, newdata, Coef_mat)
}
\arguments{
\item{Y}{a xts object, endogeneous variables}

\item{newdata}{a xts object, exogenous variables for the prediction period}

\item{Coef_mat}{a matrix, coefficients matrix}
}
\description{
Simple point predicition with OLS parameter estimates
}
