% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayesVAR.R
\name{VAR_bayes}
\alias{VAR_bayes}
\title{Bayesian VAR model
implement Bayesian VAR methods}
\usage{
VAR_bayes(
  data,
  p,
  exos = colnames(data)[ncol(data)],
  N = 1500,
  warmup = 500,
  prior = "Minnetota-Wishart",
  minnesota_par = list(rho = 1, lambdas = c(0.1, 0.5, 2, 100))
)
}
\arguments{
\item{data}{time-series data with a xts format}

\item{p}{a integer, time lags of endogenous variables}

\item{exos}{a character vector, exogenous variables}

\item{N}{a integer, the number of iterations}

\item{warmup}{a integer, the number of iterations to discard before convergence}

\item{prior}{a character,  priors and methods to be used}

\item{minnesota_par}{a list, parameters in Minnesota prior}
}
\value{
posterior of parameters, credible bands of responses in Bayesian VAR model
}
\description{
Bayesian VAR model
implement Bayesian VAR methods
}
