% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VAR_OLS.R
\name{VAR_OLS}
\alias{VAR_OLS}
\title{Ordinary Least Square (OLS) estimation for VAR model
provide OLS estimates of VAR models}
\usage{
VAR_OLS(data, p, exos = colnames(data)[ncol(data)])
}
\arguments{
\item{data}{time-series data with a xts format}

\item{p}{a integer, time lags of endogenous variables}

\item{exos}{a character vector, exogenous variables}
}
\value{
Coef_mat: a coefficient matrix, Sigma: covariance matrix,
model: a list of two matrices: endogenous variables and explanatory variables
}
\description{
Ordinary Least Square (OLS) estimation for VAR model
provide OLS estimates of VAR models
}
