% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VAR_OLS.R
\name{impulse_response}
\alias{impulse_response}
\title{Impulse-Response function of estimated VAR model
provide an impulse-response function and its plot}
\usage{
impulse_response(model, ...)
}
\arguments{
\item{model}{a object}
}
\value{
impulse-response function and data to be used in the plot
}
\description{
Impulse-Response function of estimated VAR model
provide an impulse-response function and its plot
}
