% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayesVAR.R
\name{plot.bayesVAR}
\alias{plot.bayesVAR}
\title{Figures from Bayesian VAR model
provide figures from Bayesian VAR methods}
\usage{
\method{plot}{bayesVAR}(x, y = NULL, ncol.fig = 2)
}
\arguments{
\item{x}{Bayesian VAR model}

\item{y}{NULL}

\item{ncol.fig}{a integer, number of figures plotted in a row}
}
\description{
Figures from Bayesian VAR model
provide figures from Bayesian VAR methods
}
