% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VAR_OLS.R
\name{predict.OLS}
\alias{predict.OLS}
\title{Simple point prediction with OLS parameter estimates}
\usage{
\method{predict}{OLS}(model, newdata)
}
\arguments{
\item{model}{a OLS model object}

\item{newdata}{a xts object, exogenous variables for the prediction period}
}
\description{
Simple point prediction with OLS parameter estimates
}
