% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VAR_OLS.R
\name{VAR_OLS}
\alias{VAR_OLS}
\title{Ordinary Least Square (OLS) estimation for VAR model
provide OLS estimates of VAR models}
\usage{
VAR_OLS(data, p, exos = colnames(data)[ncol(data)], confint = 0.95)
}
\arguments{
\item{data}{time-series data with a xts format}

\item{p}{a integer, time lags of endogenous variables}

\item{exos}{a character vector, exogenous variables}

\item{confint}{level of confidence interval}
}
\value{
OLS object, a coefficient matrix, covariance matrix, fitted values etc.
}
\description{
Ordinary Least Square (OLS) estimation for VAR model
provide OLS estimates of VAR models
}
